Golden-DataSet-for-Rag-main/requirement.txt
streamlit
openai
chromadb
sentence-transformers
youtube-transcript-api
rank-bm25
numpy
pyyaml